Add files via upload
This is a very easy to use module - usually it would take 40 ~ 50 lines of code to crack and create hashes. This module shortens that down to one line - just type 'import hashcrack' into the script area or in the interactive shell. Run it, and you will immediately get a hash cracking program that can crack Md5, SHA1, SHA224, SHA256 and SHA512. Just type in the hashed text and then the type of hash it is. Hashcrack will do the rest. And if you want to encrypt something as a hash, just type 'import md5encrypt' into the script area or the interactive shell. Run it and you will get an md5 encryption program. Just type 'import sha1encrypt' for sha1 encryption, 'import sha224encrypt' for sha224 encryption, 'import sha256encrypt' for sha256 encryption, 'import sha512encrypt' for sha512 encryption, and 'import sha384encrypt' for sha384 encryption.